Moved authentication middlware to middleware folder and created wrapper method
