+++ server/routes.go
@@ -8,19 +8,13 @@ import (
 	"github.com/gorilla/mux"
 
 	ctrl "github.com/HackGT/SponsorshipPortal/controller"
-	"github.com/HackGT/SponsorshipPortal/controller/auth"
+	"github.com/HackGT/SponsorshipPortal/middleware/auth"
 	"github.com/HackGT/SponsorshipPortal/logger"
 )
 
 func loggingMiddleware(out io.Writer) mux.MiddlewareFunc {
 	return func(next http.Handler) http.Handler {
 		return handlers.LoggingHandler(out, next)
-	}
-}
-
-func authMiddleware() mux.MiddlewareFunc {
-	return func(next http.Handler) http.Handler {
-		return auth.AuthHandler(next)
 	}
 }
 
@@ -36,7 +30,6 @@ func (app *App) NewRouter() http.Handler {
 
 	// Attach logging and recovery middlewares
 	log := logger.New(app.Config)
-	r.Use(authMiddleware())
 	r.Use(loggingMiddleware(log.Writer()))
 	r.Use(handlers.RecoveryHandler(handlers.RecoveryLogger(log)))
 	return r
